Bug 1946790 - Update password-card to use --border-color-transparent. r=#credential-management-reviewers
To prevent borders from appearing on the password-card, we need to
switch out --border-color to --border-color-transparent.

In order to prevent additional whitespace at the bottom of the card
when an alert is present, we needed to move the border to a
prefers-contrast media query. However, the border radius of this alert
when in prefers-contrast or forced-colors needed to be adjusted as well.

By utilizing the previously added media query, we adjust the border
radius for the message bar via CSS variables instead of writing another
media query to change the border radius.

Differential Revision: https://phabricator.services.mozilla.com/D248969