Bug 1971624 - Add support for extension-controlled prefs in Setting config. r=#recomp-reviewers,#settings-reviewers
This patch adds support for extension-controlled prefs when using the
new setting-control/config-based settings layout.

To declare that a preference is controlled by an extension, you will use
the new `extensionPrefKey` when declaring the config object for that
setting. E.g,
`Preferences.addSetting({id: 'test', extensionPrefKey: 'test-key'})`.

The Setting class will handle asking the ExtensionSettingStore and the
AddonManager for the relevant info about the particular controlling
extension, such as ID and name. The Setting class will also handle
disabling the controlling extension.

The instance of SettingControl will handle displaying the message bar
that displays which extension is controlling which pref, as well as
the button to disable said extension.