Add locale options to the add available element
Still a bit to be done though. Need to figure out how to extract
the shared logic of languages.js to a module so I can avoid the
onLoad handler of that file. It causes many cascading errors, so
isolating that from the new layout seems like the better option.